DAOS-8055 utils: Separate daos_perf and vos_perf
To get rid of server module dependency from daos_perf, the original
daos_perf is split into two binaries:

- daos_perf : for original 'daos' and 'echo' modes only.
- vos_perf  : for original 'vos' mode.

dts lib is also modified to get rid of dependency on vos, and
daos_racer is now independent of vos as well.
