+++ src/common/SConscript
@@ -41,7 +41,7 @@ def build_dts_library(env):
             print("Install and load mpich or openmpi\n")
             return
 
-    libraries = ['cmocka', 'daos_tests', 'vos', 'bio', 'abt']
+    libraries = ['cmocka', 'daos_tests', 'abt']
 
     # Add runtime paths for daos libraries
     denv.AppendUnique(LINKFLAGS=[r'-Wl,-rpath=\$$ORIGIN/daos_srv'])
