+++ src/common/dts.c
@@ -24,7 +24,6 @@
 #include <daos/common.h>
 #include <daos/credit.h>
 #include <daos/tests_lib.h>
-#include <daos_srv/vos.h>
 #include <daos_test.h>
 #include <daos/dts.h>
 
@@ -41,46 +40,13 @@ enum {
 };
 
 static int
-pool_init(struct credit_context *tsc)
+engine_pool_init(struct credit_context *tsc)
 {
 	daos_handle_t	poh = DAOS_HDL_INVAL;
-	int		rc;
+	int		rc = 0;
 
-	if (tsc->tsc_scm_size == 0)
-		tsc->tsc_scm_size = (1ULL << 30);
-
-	if (tsc->tsc_pmem_file) { /* VOS mode */
-		char	*pmem_file = tsc->tsc_pmem_file;
-		int	 fd;
-
-		if (!daos_file_is_dax(pmem_file)) {
-			rc = open(pmem_file, O_CREAT | O_TRUNC | O_RDWR, 0666);
-			if (rc < 0)
-				goto out;
-
-			fd = rc;
-			rc = fallocate(fd, 0, 0, tsc->tsc_scm_size);
-			if (rc)
-				goto out;
-		}
-		if (tsc_create_pool(tsc)) {
-			/* Use pool size as blob size for this moment. */
-			rc = vos_pool_create(pmem_file, tsc->tsc_pool_uuid, 0,
-					     tsc->tsc_nvme_size, 0, &poh);
-			if (rc)
-				goto out;
-		} else {
-			rc = vos_pool_open(pmem_file, tsc->tsc_pool_uuid, 0,
-					   &poh);
-			if (rc)
-				goto out;
-		}
-
-	} else if (tsc->tsc_mpi_rank == 0) { /* DAOS mode and rank zero */
+	if (tsc->tsc_mpi_rank == 0) {
 		d_rank_list_t	*svc = &tsc->tsc_svc;
-
-		if (tsc->tsc_dmg_conf)
-			dmg_config_file = tsc->tsc_dmg_conf;
 
 		if (tsc_create_pool(tsc)) {
 			rc = dmg_pool_create(dmg_config_file, geteuid(),
@@ -98,68 +64,43 @@ pool_init(struct credit_context *tsc)
 			goto bcast;
 	}
 	tsc->tsc_poh = poh;
- bcast:
+bcast:
 	if (tsc->tsc_mpi_size <= 1)
-		goto out; /* don't need to share handle */
+		return rc; /* don't need to share handle */
 
-	if (!tsc->tsc_pmem_file)
-		MPI_Bcast(&rc, 1, MPI_INT, 0, MPI_COMM_WORLD);
+	MPI_Bcast(&rc, 1, MPI_INT, 0, MPI_COMM_WORLD);
 	if (rc)
-		goto out; /* open failed */
+		return rc; /* open failed */
 
-	if (!tsc->tsc_pmem_file)
-		handle_share(&tsc->tsc_poh, HANDLE_POOL, tsc->tsc_mpi_rank,
-			     tsc->tsc_poh, 0);
- out:
+	handle_share(&tsc->tsc_poh, HANDLE_POOL, tsc->tsc_mpi_rank,
+		     tsc->tsc_poh, 0);
 	return rc;
 }
 
 static void
-pool_fini(struct credit_context *tsc)
+engine_pool_fini(struct credit_context *tsc)
 {
 	int	rc;
 
-	if (tsc->tsc_pmem_file) { /* VOS mode */
-		vos_pool_close(tsc->tsc_poh);
-		if (tsc_create_pool(tsc)) {
-			rc = vos_pool_destroy(tsc->tsc_pmem_file,
-					      tsc->tsc_pool_uuid);
-			D_ASSERTF(rc == 0 || rc == -DER_NONEXIST,
-				  "rc: "DF_RC"\n", DP_RC(rc));
-		}
-	} else { /* DAOS mode */
-		rc = daos_pool_disconnect(tsc->tsc_poh, NULL);
-		D_ASSERTF(rc == 0 || rc == -DER_NO_HDL, "rc="DF_RC"\n",
-			  DP_RC(rc));
-		MPI_Barrier(MPI_COMM_WORLD);
-		if (tsc->tsc_mpi_rank == 0 && tsc_create_pool(tsc)) {
-			rc = dmg_pool_destroy(dmg_config_file,
-					      tsc->tsc_pool_uuid, NULL, true);
-			D_ASSERTF(rc == 0 || rc == -DER_NONEXIST ||
-				  rc == -DER_TIMEDOUT, "rc="DF_RC"\n",
-				  DP_RC(rc));
-		}
+	rc = daos_pool_disconnect(tsc->tsc_poh, NULL);
+	D_ASSERTF(rc == 0 || rc == -DER_NO_HDL, "rc="DF_RC"\n", DP_RC(rc));
+	MPI_Barrier(MPI_COMM_WORLD);
+
+	if (tsc->tsc_mpi_rank == 0 && tsc_create_pool(tsc)) {
+		rc = dmg_pool_destroy(dmg_config_file, tsc->tsc_pool_uuid,
+				      NULL, true);
+		D_ASSERTF(rc == 0 || rc == -DER_NONEXIST ||
+			  rc == -DER_TIMEDOUT, "rc="DF_RC"\n", DP_RC(rc));
 	}
 }
 
 static int
-cont_init(struct credit_context *tsc)
+engine_cont_init(struct credit_context *tsc)
 {
 	daos_handle_t	coh = DAOS_HDL_INVAL;
-	int		rc;
+	int		rc = 0;
 
-	if (tsc->tsc_pmem_file) { /* VOS mode */
-		if (tsc_create_cont(tsc)) {
-			rc = vos_cont_create(tsc->tsc_poh, tsc->tsc_cont_uuid);
-			if (rc)
-				goto out;
-		}
-
-		rc = vos_cont_open(tsc->tsc_poh, tsc->tsc_cont_uuid, &coh);
-		if (rc)
-			goto out;
-
-	} else if (tsc->tsc_mpi_rank == 0) { /* DAOS mode and rank zero */
+	if (tsc->tsc_mpi_rank == 0) {
 		if (tsc_create_cont(tsc)) {
 			rc = daos_cont_create(tsc->tsc_poh, tsc->tsc_cont_uuid,
 					      NULL, NULL);
@@ -172,34 +113,50 @@ cont_init(struct credit_context *tsc)
 			goto bcast;
 	}
 	tsc->tsc_coh = coh;
- bcast:
+bcast:
 	if (tsc->tsc_mpi_size <= 1)
-		goto out; /* don't need to share handle */
+		return rc; /* don't need to share handle */
 
 	MPI_Bcast(&rc, 1, MPI_INT, 0, MPI_COMM_WORLD);
 	if (rc)
-		goto out; /* open failed */
+		return rc; /* open failed */
 
-	if (!tsc->tsc_pmem_file)
-		handle_share(&tsc->tsc_coh, HANDLE_CO, tsc->tsc_mpi_rank,
-			     tsc->tsc_poh, 0);
- out:
+	handle_share(&tsc->tsc_coh, HANDLE_CO, tsc->tsc_mpi_rank,
+		     tsc->tsc_poh, 0);
 	return rc;
 }
 
 static void
-cont_fini(struct credit_context *tsc)
+engine_cont_fini(struct credit_context *tsc)
 {
-	if (tsc->tsc_pmem_file) /* VOS mode */
-		vos_cont_close(tsc->tsc_coh);
-	else /* DAOS mode */
-		daos_cont_close(tsc->tsc_coh, NULL);
-
+	daos_cont_close(tsc->tsc_coh, NULL);
 	/* NB: no container destroy at here, it will be destroyed by pool
 	 * destroy later. This is because container destroy could be too
 	 * expensive after performance tests.
 	 */
 }
+
+static void
+engine_fini(void)
+{
+	daos_fini();
+}
+
+static int
+engine_init(void)
+{
+	return daos_init();
+}
+
+struct io_engine daos_engine = {
+	.ie_name	= "DAOS",
+	.ie_init	= engine_init,
+	.ie_fini	= engine_fini,
+	.ie_pool_init	= engine_pool_init,
+	.ie_pool_fini	= engine_pool_fini,
+	.ie_cont_init	= engine_cont_init,
+	.ie_cont_fini	= engine_cont_fini,
+};
 
 bool
 dts_is_async(struct credit_context *tsc)
@@ -209,30 +166,38 @@ dts_is_async(struct credit_context *tsc)
 
 /* see comments in daos/dts.h */
 int
-dts_ctx_init(struct credit_context *tsc)
+dts_ctx_init(struct credit_context *tsc, struct io_engine *engine)
 {
 	int	rc;
 
 	tsc->tsc_init = DTS_INIT_NONE;
+	/* Use default 'DAOS' engine when no engine specified */
+	tsc->tsc_engine = (engine == NULL) ? &daos_engine : engine;
+
 	rc = daos_debug_init(DAOS_LOG_DEFAULT);
 	if (rc)
 		goto out;
 	tsc->tsc_init = DTS_INIT_DEBUG;
 
-	if (tsc->tsc_pmem_file) /* VOS mode */
-		rc = vos_self_init("/mnt/daos");
-	else
-		rc = daos_init();
+	D_ASSERT(tsc->tsc_engine->ie_init != NULL);
+	rc = tsc->tsc_engine->ie_init();
 	if (rc)
 		goto out;
 	tsc->tsc_init = DTS_INIT_MODULE;
 
-	rc = pool_init(tsc);
+	if (tsc->tsc_scm_size == 0)
+		tsc->tsc_scm_size = (1ULL << 30);
+	if (tsc->tsc_dmg_conf)
+		dmg_config_file = tsc->tsc_dmg_conf;
+
+	D_ASSERT(tsc->tsc_engine->ie_pool_init != NULL);
+	rc = tsc->tsc_engine->ie_pool_init(tsc);
 	if (rc)
 		goto out;
 	tsc->tsc_init = DTS_INIT_POOL;
 
-	rc = cont_init(tsc);
+	D_ASSERT(tsc->tsc_engine->ie_cont_init != NULL);
+	rc = tsc->tsc_engine->ie_cont_init(tsc);
 	if (rc)
 		goto out;
 	tsc->tsc_init = DTS_INIT_CONT;
@@ -260,16 +225,16 @@ dts_ctx_fini(struct credit_context *tsc)
 		credits_fini(tsc);
 		/* fall through */
 	case DTS_INIT_CONT:	/* close and destroy container */
-		cont_fini(tsc);
+		D_ASSERT(tsc->tsc_engine->ie_cont_fini != NULL);
+		tsc->tsc_engine->ie_cont_fini(tsc);
 		/* fall through */
 	case DTS_INIT_POOL:	/* close and destroy pool */
-		pool_fini(tsc);
+		D_ASSERT(tsc->tsc_engine->ie_pool_fini != NULL);
+		tsc->tsc_engine->ie_pool_fini(tsc);
 		/* fall through */
 	case DTS_INIT_MODULE:	/* finalize module */
-		if (tsc->tsc_pmem_file)
-			vos_self_fini();
-		else
-			daos_fini();
+		D_ASSERT(tsc->tsc_engine->ie_fini != NULL);
+		tsc->tsc_engine->ie_fini();
 		/* fall through */
 	case DTS_INIT_DEBUG:	/* finalize debug system */
 		daos_debug_fini();
