+++ src/tests/SConscript
@@ -29,21 +29,31 @@ def scons():
 
     daos_build.program(denv, 'simple_array', 'simple_array.c', LIBS=libs)
     daos_build.program(denv, 'simple_obj', 'simple_obj.c', LIBS=libs)
-    libs += ['vos', 'bio', 'pthread', 'abt', 'dts']
-
-    daos_perf = daos_build.program(denv, 'daos_perf',
-                                   ['daos_perf.c'],
-                                   LIBS=libs)
-    denv.Install('$PREFIX/bin/', daos_perf)
+    libs += ['pthread', 'abt', 'dts']
 
     daos_racer = daos_build.program(denv, 'daos_racer',
                                     ['daos_racer.c'],
                                     LIBS=libs)
     denv.Install('$PREFIX/bin/', daos_racer)
 
+    perf_common = denv.SharedObject(['perf_common.c'])
+
+    daos_perf = daos_build.program(denv, 'daos_perf',
+                                   ['daos_perf.c', perf_common],
+                                   LIBS = libs)
+    denv.Install('$PREFIX/bin/', daos_perf)
+
+    libs += ['vos', 'bio', 'abt']
+    vos_engine = denv.SharedObject(['vos_engine.c'])
+
+    vos_perf = daos_build.program(denv, 'vos_perf',
+                                  ['vos_perf.c', perf_common, vos_engine],
+                                  LIBS = libs)
+    denv.Install('$PREFIX/bin/', vos_perf)
+
     Import('mpi_cmd_parser')
     obj_ctl = daos_build.program(denv, 'obj_ctl',
-                                 ['obj_ctl.c', mpi_cmd_parser],
+                                 ['obj_ctl.c', mpi_cmd_parser, vos_engine],
                                  LIBS=libs)
     denv.Install('$PREFIX/bin/', obj_ctl)
