+++ src/tests/obj_ctl.c
@@ -370,6 +370,8 @@ static struct option ctl_ops[] = {
 	{ NULL,		0,			NULL,	0	},
 };
 
+extern struct io_engine vos_engine;
+
 int
 main(int argc, char *argv[])
 {
@@ -395,7 +397,7 @@ main(int argc, char *argv[])
 
 	ctl_ctx.tsc_pmem_file = pmem_file;
 
-	rc = dts_ctx_init(&ctl_ctx);
+	rc = dts_ctx_init(&ctl_ctx, &vos_engine);
 	if (rc != 0) {
 		fprintf(stderr, "Failed to initialize utility: "DF_RC"\n",
 			DP_RC(rc));
